+++ electron/preload.js
@@ -50,6 +50,7 @@ contextBridge.exposeInMainWorld('electronAPI', {
   chooseReportScope: () => ipcRenderer.invoke('choose-report-scope'),
   setExportReportAvailable: (enabled) => ipcRenderer.send('set-export-report-available', enabled),
   openVideo: (filePath) => ipcRenderer.invoke('open-video', filePath),
+  setVideoFullscreen: (fullscreen) => ipcRenderer.invoke('set-video-fullscreen', fullscreen),
 
   // Menu events
   onMenuAction: (callback) => {